requirements.txt
pytest==7.4.0
requests==2.32.5
allure-pytest==2.15.0